Revert "Refactored the package name for OSS"
This reverts commit e81d187a6944143b9d26a16e1e4082532e635b90.
